# backend/docs/swagger.yaml
openapi: 3.0.0
info:
  title: API de Pessoas e Assinaturas
  version: 1.0.0
paths:
  /api/alterar/{cpf}:
    put:
      summary: Atualiza os dados de uma pessoa pelo CPF
      tags:
        - Pessoas
      parameters:
        - in: path
          name: cpf
          required: true
          schema:
            type: string
          description: CPF da pessoa
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nome:
                  type: string
                rua:
                  type: string
                numero:
                  type: integer
                bairro:
                  type: string
              example:
                nome: "Carlos Silva"
                rua: "Avenida Brasil"
                numero: 456
                bairro: "Jardins"
      responses:
        200:
          description: Dados atualizados com sucesso
        404:
          description: Pessoa não encontrada

  /api/buscar/{cpf}:
    get:
      summary: Busca uma pessoa pelo CPF e suas assinaturas
      tags:
        - Pessoas
      parameters:
        - in: path
          name: cpf
          required: true
          schema:
            type: string
          description: CPF da pessoa
      responses:
        200:
          description: Pessoa encontrada, incluindo suas assinaturas
          content:
            application/json:
              schema:
                type: object
                properties:
                  pessoa:
                    $ref: '#/components/schemas/Pessoa'
                  assinaturas:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                        pessoaCpf:
                          type: string
                        dataAssinatura:
                          type: string
                          format: date
                        quantidade:
                          type: integer
        404:
          description: Pessoa não encontrada
        500:
          description: Erro ao buscar dados

  /api/novo:
    post:
      summary: Cria uma nova Pessoa (CPF)
      tags:
        - Pessoas
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                cpf:
                  type: string
                nome:
                  type: string
                rua:
                  type: string
                numero:
                  type: integer
                bairro:
                  type: string
              example:
                cpf: "12345678900"
                nome: "Carlos Silva"
                rua: "Avenida Brasil"
                numero: 345
                bairro: "Jardins"
      responses:
        201:
          description: Pessoa criada com sucesso
        400:
          description: Erro ao criar Pessoa (CPF já existe)

  /api/listar:
    get:
      summary: Lista todas as Pessoas cadastradas
      tags:
        - Pessoas
      responses:
        200:
          description: Lista de Pessoas retornada com sucesso
        500:
          description: Erro ao buscar pessoas

  /api/assinatura:
    post:
      summary: Adiciona uma nova assinatura para uma pessoa
      tags:
        - Assinaturas
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - CPF da pessoa
                - Data da Assinatura
                - Quantidade
              properties:
                CPF da pessoa:
                  type: string
                  description: CPF da pessoa vinculada à assinatura
                Data da Assinatura:
                  type: string
                  format: date
                  description: Data da assinatura no formato DD-MM-AAAA
                Quantidade:
                  type: integer
                  description: Quantidade de itens na assinatura
      responses:
        201:
          description: Assinatura criada com sucesso
        400:
          description: CPF não cadastrado ou dados inválidos
        500:
          description: Erro ao adicionar assinatura

  /api/assinatura/{cpf}:
    get:
      summary: Lista todas as assinaturas de uma pessoa pelo CPF
      tags:
        - Assinaturas
      parameters:
        - in: path
          name: cpf
          required: true
          schema:
            type: string
          description: CPF da pessoa para buscar as assinaturas
      responses:
        200:
          description: Lista de assinaturas da pessoa
        404:
          description: Nenhuma assinatura encontrada (Banco vazio)
        500:
          description: Erro ao buscar assinaturas